dashboard/serializers/states_districts.py
from rest_framework import serializers
from dashboard.models import State, District


class StateSerializer(serializers.ModelSerializer):
    class Meta:
        model = State
        fields = '__all__'


class DistrictSerializer(serializers.ModelSerializer):
    class Meta:
        model = District
        fields = '__all__'
